Implement daily insights and overhaul analytics UI
### Core Changes
- **New `DailyInsight` System**: Introduced a cross-platform caching layer (`DailyInsightCache`) and notification handler to receive and display daily study tips or insights.
- **Analytics & Tracking**:
    - Added `sessionCompletionRate` and `responseTimeTrend` to study insights.
    - Updated `wordId` from `Int` to `Long` across all analytics models for consistency with the database.
    - Enhanced `AnalyticsRecorderImpl` with a retry queue to handle failed session synchronizations on subsequent attempts.
- **Data Layer Improvements**: Added remote and repository support for fetching level transitions and response time trends.

### UI & Experience
- **Redesigned Insights Screen**:
    - Replaced tabbed navigation with a single-scrollable view containing Overview, Trends, and Words sections.
    - Added a "Hero" section showing total cards reviewed and a new "This Week" activity bar chart.
    - Redesigned stat cards and metric rows with a cleaner, border-based aesthetic.
    - Integrated a dismissible `DailyInsightBanner` for the new daily insights feature.
- **Study Screen**: Added an "Insights" action button to the top bar for quick access to analytics.
- **Profile Screen**: Removed the redundant `WeeklyActivitySection` in favor of the consolidated insights view.

### Technical Details
- Added `GetLevelTransitionsUseCase` and `GetResponseTimeTrendUseCase`.
- Implemented platform-specific `DailyInsightCache` for Android (`SharedPreferences`), iOS (`NSUserDefaults`), and Wasm (in-memory).
- Refactored `NotificationPayloadHandlerRegistry` to pass notification body text to handlers.

+++ domain/src/commonMain/kotlin/domain/analytics/model/ComebackWord.kt
@@ -1,7 +1,7 @@
 package domain.analytics.model
 
 data class ComebackWord(
-    val wordId: Int,
+    val wordId: Long,
     val wordText: String,
     val wordTranslation: String,
 )
