+++ domain/src/commonMain/kotlin/domain/analytics/model/MasteredWord.kt
@@ -1,7 +1,7 @@
 package domain.analytics.model
 
 data class MasteredWord(
-    val wordId: Int,
+    val wordId: Long,
     val wordText: String,
     val wordTranslation: String,
     val masteredAt: Long,
