+++ domain/src/commonMain/kotlin/domain/analytics/model/MostReviewedWord.kt
@@ -1,7 +1,7 @@
 package domain.analytics.model
 
 data class MostReviewedWord(
-    val wordId: Int,
+    val wordId: Long,
     val wordText: String,
     val wordTranslation: String,
     val totalReviews: Int,
